refactor(trigger): reuse shared ssrfGuard instead of inline SSRF logic
Replaces the ~130-line inline SSRF implementation (isPrivateIp,
validateCallbackUrl, safeFetchCallback) with imports from the new
shared utils/ssrfGuard.js module. Behaviour is identical.
